requirements.txt
alembic==1.12.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
coverage==7.3.1
ecdsa==0.18.0
fastapi==0.103.1
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pyasn1==0.5.0
pydantic==1.10.12
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
strawberry-graphql==0.205.0
typing_extensions==4.7.1
uvicorn==0.23.2
